(PC-17806) feat(auth): move useNextSubscriptionStep to api folder
