+++ src/features/home/components/headers/HomeHeader.tsx
@@ -3,9 +3,9 @@ import React, { FunctionComponent, useMemo } from 'react'
 import { Platform } from 'react-native'
 import styled from 'styled-components/native'
 
+import { useNextSubscriptionStep } from 'features/auth/api/useNextSubscriptionStep'
 import { useAuthContext } from 'features/auth/context/AuthContext'
 import { useBeneficiaryValidationNavigation } from 'features/auth/helpers/useBeneficiaryValidationNavigation'
-import { useNextSubscriptionStep } from 'features/auth/helpers/useNextSubscriptionStep'
 import { GeolocationBanner } from 'features/home/components/GeolocationBanner'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { isUserBeneficiary } from 'features/profile/helpers/isUserBeneficiary'
